+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v0.12.3 - WIP
+* Get rid of Ext's viewport warning
 
 # v0.12.2 - 2015-06-06
 * Fix loading multi-instance components